+++ lib/TraceState.ts
@@ -11,12 +11,22 @@ export class TraceState {
     return Object.keys(this.internalState);
   }
 
-  get(name: string): string {
+  get(name: string): string | undefined {
     return this.internalState[name];
   }
 
   set(name: string, value: string): void {
-    this.internalState[name] = value;
+    this.internalState = {
+      // ensure that the new key ends up in the beginning of the list
+      [name]: value,
+      ...this.internalState,
+      // ensure that updates work
+      [name]: value
+    };
+  }
+
+  getInternalState(): InternalTraceState {
+    return this.internalState;
   }
 
   clone() {