+++ lib/__test__/TraceContext.test.ts
@@ -1,22 +1,22 @@
-import {TraceContext} from '../TraceContext';
+import {TraceParent} from '../TraceParent';
 import {TraceState} from '../TraceState';
 
-describe('TraceContext', () => {
+describe('TraceParent', () => {
   describe('flags', () => {
     it('must identify whether tracing is enabled', () => {
-      expect(getTraceContextWithOptions(1).isTracedFlagSet()).toEqual(true);
-      expect(getTraceContextWithOptions(0).isTracedFlagSet()).toEqual(false);
+      expect(getTraceParentWithOptions(1).isTracedFlagSet()).toEqual(true);
+      expect(getTraceParentWithOptions(0).isTracedFlagSet()).toEqual(false);
     });
 
     it('must identify other flags as well', () => {
-      expect(getTraceContextWithOptions(2).isFlagSet(1)).toEqual(false);
-      expect(getTraceContextWithOptions(3).isFlagSet(4)).toEqual(false);
-      expect(getTraceContextWithOptions(3).isFlagSet(1)).toEqual(true);
-      expect(getTraceContextWithOptions(3).isFlagSet(2)).toEqual(true);
+      expect(getTraceParentWithOptions(2).isFlagSet(1)).toEqual(false);
+      expect(getTraceParentWithOptions(3).isFlagSet(4)).toEqual(false);
+      expect(getTraceParentWithOptions(3).isFlagSet(1)).toEqual(true);
+      expect(getTraceParentWithOptions(3).isFlagSet(2)).toEqual(true);
     });
 
     it('must set flags', () => {
-      const ctx = getTraceContextWithOptions(0);
+      const ctx = getTraceParentWithOptions(0);
       ctx.setTracedFlag();
       expect(ctx.isFlagSet(1)).toEqual(true);
       expect(ctx.isFlagSet(2)).toEqual(false);
@@ -26,24 +26,22 @@ describe('TraceContext', () => {
       expect(ctx.isFlagSet(2)).toEqual(true);
     });
 
-    function getTraceContextWithOptions(options: number): TraceContext {
-      return new TraceContext({
+    function getTraceParentWithOptions(options: number): TraceParent {
+      return new TraceParent({
         version: 0,
         traceId: 'abc',
         spanId: 'abc',
-        options,
-        state: new TraceState({})
+        options
       });
     }
   });
 
   it('must clone', () => {
-    const input = new TraceContext({
+    const input = new TraceParent({
       version: 0,
       traceId: 'abc',
       spanId: 'abc',
-      options: 0,
-      state: new TraceState({})
+      options: 0
     });
     const clone = input.clone();
     expect(clone).not.toBe(input);