+++ lib/__test__/TraceState.test.ts
@@ -1,4 +1,3 @@
-import {TraceContext} from '../TraceContext';
 import {TraceState} from '../TraceState';
 
 describe('TraceState', () => {
@@ -12,14 +11,15 @@ describe('TraceState', () => {
   it('must add new keys', () => {
     const state = new TraceState({a: '1', b: '2'});
     state.set('c', '3');
-    expect(state.keys()).toEqual(['a', 'b', 'c']);
+    expect(state.keys()).toEqual(['c', 'a', 'b',]);
     expect(state.get('c')).toEqual('3');
   });
 
-  it('must replace keys', () => {
+  it('must replace keys and move them to the front', () => {
     const state = new TraceState({a: '1', b: '2'});
-    state.set('a', '3');
-    expect(state.keys()).toEqual(['a', 'b']);
-    expect(state.get('a')).toEqual('3');
+    state.set('b', '3');
+    expect(state.keys()).toEqual(['b', 'a']);
+    expect(state.get('a')).toEqual('1');
+    expect(state.get('b')).toEqual('3');
   });
 });