+++ lib/format/http/index.ts
@@ -1,0 +1,13 @@
+export {
+  extract as extractTraceParent,
+  inject as injectTraceParent,
+  parse as parseTraceParent,
+  serialize as serializeTraceParent
+} from './traceParent';
+
+export {
+  extract as extractTraceState,
+  inject as injectTraceState,
+  parse as parseTraceState,
+  serialize as serializeTraceState
+} from './traceState';